tests/wpt/tests/css/css-conditional/container-queries/animation-container-size.html

<!doctype html>
<title>Container Queries - Animating container size</title>
<link rel="help" href="https://drafts.csswg.org/css-conditional-5/#animated-containers">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/cq-testcommon.js"></script>
<style>
  @keyframes anim {
    from { width: 200px; }
    to { width: 100px; }
  }
  #container {
    container-type: inline-size;
    animation: anim 1s linear paused;
  }
  #target {
    background-color: green;
  }

  @container (width: 200px) {
    #target {
      background-color: blue;
    }
  }
</style>
<div id=container>
  <div id=target>
    Test
  </div>
</div>
<script>
  setup(() => assert_implements_size_container_queries());

  test(() => {
    assert_equals(getComputedStyle(target).backgroundColor, 'rgb(0, 0, 255)');

    assert_equals(container.getAnimations().length, 1);
    let animation = container.getAnimations()[0];
    animation.currentTime = 500;

    assert_equals(getComputedStyle(target).backgroundColor, 'rgb(0, 128, 0)');
  }, 'Animation affects container query evaluation');
</script>
